+++ sdk/storage/storage-blob/recordings/node/blobclient_nodejs_only/recording_setmetadata_with_cleaning_up_metadata.js
@@ -1,107 +1,119 @@
 let nock = require('nock');
 
-module.exports.testInfo = {"container":"container156816861121402747","blob":"blob156816861162106037"}
+module.exports.hash = "5fcb62eeaa74c436ad1c241b97d08805";
+
+module.exports.testInfo = {"uniqueName":{"container":"container159210827790005161","blob":"blob159210827791505055"},"newDate":{}}
 
 nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/container156816861121402747')
+  .put('/container159210827790005161')
   .query(true)
-  .reply(201, "", [ 'Content-Length',
+  .reply(201, "", [
+  'Content-Length',
   '0',
   'Last-Modified',
-  'Wed, 11 Sep 2019 02:23:31 GMT',
+  'Sun, 14 Jun 2020 04:17:57 GMT',
   'ETag',
-  '"0x8D7365F0A19FE5F"',
+  '"0x8D81019EB322F93"',
   'Server',
   'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '1837c8b1-b01e-004d-0747-68a01e000000',
+  'c1309795-201e-003e-5a02-42dadf000000',
   'x-ms-client-request-id',
-  '137b33d5-c06f-47ca-b9b4-dbf97b82573b',
+  '17b4f5a7-fdf0-4713-9aea-850063ca4871',
   'x-ms-version',
-  '2019-02-02',
+  '2019-12-12',
   'Date',
-  'Wed, 11 Sep 2019 02:23:31 GMT' ]);
-
+  'Sun, 14 Jun 2020 04:17:56 GMT'
+]);
 
 nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/container156816861121402747/blob156816861162106037', "Hello World")
-  .reply(201, "", [ 'Content-Length',
+  .put('/container159210827790005161/blob159210827791505055', "Hello World")
+  .reply(201, "", [
+  'Content-Length',
   '0',
   'Content-MD5',
   'sQqNsWTgdUEFt6mb5y4/5Q==',
   'Last-Modified',
-  'Wed, 11 Sep 2019 02:23:31 GMT',
+  'Sun, 14 Jun 2020 04:17:57 GMT',
   'ETag',
-  '"0x8D7365F0A59DEAC"',
+  '"0x8D81019EB34BD25"',
   'Server',
   'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'e18b4285-d01e-0030-6347-683cd6000000',
+  'c130979b-201e-003e-5f02-42dadf000000',
   'x-ms-client-request-id',
-  '374350d4-f868-40aa-b2a6-ee1ad5af7cef',
+  '44c3b451-e7a7-435b-8313-02c308e0f2ef',
   'x-ms-version',
-  '2019-02-02',
+  '2019-12-12',
   'x-ms-content-crc64',
   'YeJLfssylmU=',
   'x-ms-request-server-encrypted',
   'true',
+  'x-ms-version-id',
+  '2020-06-14T04:17:57.9264060Z',
   'Date',
-  'Wed, 11 Sep 2019 02:23:31 GMT' ]);
-
+  'Sun, 14 Jun 2020 04:17:56 GMT'
+]);
 
 nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/container156816861121402747/blob156816861162106037')
+  .put('/container159210827790005161/blob159210827791505055')
   .query(true)
-  .reply(200, "", [ 'Content-Length',
+  .reply(200, "", [
+  'Content-Length',
   '0',
   'Last-Modified',
-  'Wed, 11 Sep 2019 02:23:32 GMT',
+  'Sun, 14 Jun 2020 04:17:57 GMT',
   'ETag',
-  '"0x8D7365F0A98EEE0"',
+  '"0x8D81019EB372E96"',
   'Server',
   'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '84b8f259-401e-0017-5e47-68a69f000000',
+  'c13097a1-201e-003e-6502-42dadf000000',
   'x-ms-client-request-id',
-  '77edba80-3fcc-4838-9959-17fc00b164ab',
+  '3d17bcbd-bc7a-43e6-99ba-dab4b8ef3084',
   'x-ms-version',
-  '2019-02-02',
+  '2019-12-12',
+  'x-ms-version-id',
+  '2020-06-14T04:17:57.9424166Z',
   'x-ms-request-server-encrypted',
   'true',
   'Date',
-  'Wed, 11 Sep 2019 02:23:31 GMT' ]);
-
+  'Sun, 14 Jun 2020 04:17:56 GMT'
+]);
 
 nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
-  .head('/container156816861121402747/blob156816861162106037')
-  .reply(200, "", [ 'Content-Length',
+  .head('/container159210827790005161/blob159210827791505055')
+  .reply(200, "", [
+  'Content-Length',
   '11',
   'Content-Type',
   'application/octet-stream',
   'Content-MD5',
   'sQqNsWTgdUEFt6mb5y4/5Q==',
   'Last-Modified',
-  'Wed, 11 Sep 2019 02:23:32 GMT',
+  'Sun, 14 Jun 2020 04:17:57 GMT',
   'Accept-Ranges',
   'bytes',
   'ETag',
-  '"0x8D7365F0A98EEE0"',
+  '"0x8D81019EB372E96"',
   'Server',
   'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'f7be116f-301e-005c-1147-689705000000',
+  'c13097a7-201e-003e-6b02-42dadf000000',
   'x-ms-client-request-id',
-  '07489006-5458-43e2-8e2e-868af5c4a752',
+  'a175e48e-7fc8-4326-ba5d-238d82e88902',
   'x-ms-version',
-  '2019-02-02',
-  'x-ms-tag-count',
-  '0',
+  '2019-12-12',
+  'x-ms-version-id',
+  '2020-06-14T04:17:57.9424166Z',
+  'x-ms-is-current-version',
+  'true',
   'x-ms-meta-a',
   'a',
   'x-ms-meta-b',
   'b',
   'x-ms-creation-time',
-  'Wed, 11 Sep 2019 02:23:31 GMT',
+  'Sun, 14 Jun 2020 04:17:57 GMT',
   'x-ms-lease-status',
   'unlocked',
   'x-ms-lease-state',
@@ -111,66 +123,68 @@ nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParam
   'x-ms-server-encrypted',
   'true',
   'x-ms-access-tier',
-  'Cool',
+  'Hot',
   'x-ms-access-tier-inferred',
   'true',
-  'Access-Control-Expose-Headers',
-  'x-ms-request-id,x-ms-client-request-id,Server,x-ms-version,x-ms-tag-count,x-ms-meta-a,x-ms-meta-b,Content-Type,Last-Modified,ETag,x-ms-creation-time,Content-MD5,x-ms-lease-status,x-ms-lease-state,x-ms-blob-type,x-ms-server-encrypted,x-ms-access-tier,x-ms-access-tier-inferred,Accept-Ranges,Content-Length,Date,Transfer-Encoding',
-  'Access-Control-Allow-Origin',
-  '*',
   'Date',
-  'Wed, 11 Sep 2019 02:23:31 GMT' ]);
-
+  'Sun, 14 Jun 2020 04:17:56 GMT'
+]);
 
 nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/container156816861121402747/blob156816861162106037')
+  .put('/container159210827790005161/blob159210827791505055')
   .query(true)
-  .reply(200, "", [ 'Content-Length',
+  .reply(200, "", [
+  'Content-Length',
   '0',
   'Last-Modified',
-  'Wed, 11 Sep 2019 02:23:33 GMT',
+  'Sun, 14 Jun 2020 04:17:57 GMT',
   'ETag',
-  '"0x8D7365F0B133E01"',
+  '"0x8D81019EB3AFFD6"',
   'Server',
   'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '1ab5b2fb-a01e-003f-6c47-68d120000000',
+  'c13097a9-201e-003e-6d02-42dadf000000',
   'x-ms-client-request-id',
-  '29aa42af-79f1-4189-9aa7-81f0e3cff61d',
+  '7e888835-f900-417d-8bd5-0da0a7059cb5',
   'x-ms-version',
-  '2019-02-02',
+  '2019-12-12',
+  'x-ms-version-id',
+  '2020-06-14T04:17:57.9674342Z',
   'x-ms-request-server-encrypted',
   'true',
   'Date',
-  'Wed, 11 Sep 2019 02:23:32 GMT' ]);
-
+  'Sun, 14 Jun 2020 04:17:57 GMT'
+]);
 
 nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
-  .head('/container156816861121402747/blob156816861162106037')
-  .reply(200, "", [ 'Content-Length',
+  .head('/container159210827790005161/blob159210827791505055')
+  .reply(200, "", [
+  'Content-Length',
   '11',
   'Content-Type',
   'application/octet-stream',
   'Content-MD5',
   'sQqNsWTgdUEFt6mb5y4/5Q==',
   'Last-Modified',
-  'Wed, 11 Sep 2019 02:23:33 GMT',
+  'Sun, 14 Jun 2020 04:17:57 GMT',
   'Accept-Ranges',
   'bytes',
   'ETag',
-  '"0x8D7365F0B133E01"',
+  '"0x8D81019EB3AFFD6"',
   'Server',
   'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '478ecaa6-701e-0014-7b47-68a598000000',
+  'c13097ac-201e-003e-7002-42dadf000000',
   'x-ms-client-request-id',
-  '3f1f2bca-3050-43f6-98a0-d74f24dddce0',
+  'efaf3045-55ac-432c-9b84-d0445fcf9d8c',
   'x-ms-version',
-  '2019-02-02',
-  'x-ms-tag-count',
-  '0',
+  '2019-12-12',
+  'x-ms-version-id',
+  '2020-06-14T04:17:57.9674342Z',
+  'x-ms-is-current-version',
+  'true',
   'x-ms-creation-time',
-  'Wed, 11 Sep 2019 02:23:31 GMT',
+  'Sun, 14 Jun 2020 04:17:57 GMT',
   'x-ms-lease-status',
   'unlocked',
   'x-ms-lease-state',
@@ -180,30 +194,27 @@ nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParam
   'x-ms-server-encrypted',
   'true',
   'x-ms-access-tier',
-  'Cool',
+  'Hot',
   'x-ms-access-tier-inferred',
   'true',
-  'Access-Control-Expose-Headers',
-  'x-ms-request-id,x-ms-client-request-id,Server,x-ms-version,x-ms-tag-count,Content-Type,Last-Modified,ETag,x-ms-creation-time,Content-MD5,x-ms-lease-status,x-ms-lease-state,x-ms-blob-type,x-ms-server-encrypted,x-ms-access-tier,x-ms-access-tier-inferred,Accept-Ranges,Content-Length,Date,Transfer-Encoding',
-  'Access-Control-Allow-Origin',
-  '*',
   'Date',
-  'Wed, 11 Sep 2019 02:23:33 GMT' ]);
-
+  'Sun, 14 Jun 2020 04:17:57 GMT'
+]);
 
 nock('https://fakestorageaccount.blob.core.windows.net:443', {"encodedQueryParams":true})
-  .delete('/container156816861121402747')
+  .delete('/container159210827790005161')
   .query(true)
-  .reply(202, "", [ 'Content-Length',
+  .reply(202, "", [
+  'Content-Length',
   '0',
   'Server',
   'Windows-Azure-Blob/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '42fb9f5c-101e-0062-2947-682124000000',
+  'c13097ad-201e-003e-7102-42dadf000000',
   'x-ms-client-request-id',
-  '70a67100-c6af-4e88-b3df-5db52d15f5a4',
+  '872c81c1-f9f6-496b-889d-e2326b9f6071',
   'x-ms-version',
-  '2019-02-02',
+  '2019-12-12',
   'Date',
-  'Wed, 11 Sep 2019 02:23:33 GMT' ]);
-
+  'Sun, 14 Jun 2020 04:17:57 GMT'
+]);